requirements.txt
agentscope>=0.1.0
litellm>=1.0.0
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
numpy>=1.24
tenacity>=8.0
aiofiles>=23.0
streamlit>=1.32.0
pandas>=2.0.0
